requirements.txt
fastapi==0.111.0
uvicorn==0.23.2
requests==2.32.0
pydantic==2.5.2
google-generativeai==0.1.3
